app/requirements.txt
Flask
pillow
ollama